resolve most of Paul's comments
